Move Element to its own module.
